Remove homepage link to tempest in a teapot.
It's not frontpage news.  For my own safety and, potentially, the safety
of others, I believed it wise to feature my response prominently.  It
can now be de-emphasized.

The linked page will, of course, stay right where it is.  TODO:  Add a
proper index or sitemap, with links to all things big and small.